feat: frontend 배포 관련 파일 추가 및 수정
