+++ .yarnrc.yml
@@ -1,1 +1,2 @@
 yarnPath: .yarn/releases/yarn-4.1.1.cjs
+# nodeLinker: node-modules # 배포 때 사용
