+++ src/styles/MONITORING.css
@@ -323,7 +323,7 @@
 }
 
 .chart-container {
-    background-color: #2e3240;
+    background-color: rgb(56, 60, 74);
     border: 1px solid #abafbd;
     border-radius: 8px;
     padding: 20px 20px 40px 20px;
