+++ vite.config.js
@@ -3,5 +3,9 @@ import react from '@vitejs/plugin-react-swc'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  server: {
+    host: '0.0.0.0',
+    port: 5173,
+  },
+plugins: [react()],
 })
